+++ src/patterns/rule.ts
@@ -306,23 +306,14 @@ function cacheTest(rs: null | PatternListCompiled, path: string): string | null 
 	if (rs.list.length === 1) return rs.list[0]!
 
 	let items = rs.compiledItems
-	if (!items && rs.patternSources) {
-		const len = rs.list.length
-		items = new Array(len)
-		const { nocase, patternSources: sources } = rs
-		const flags = nocase ? "i" : ""
-		for (let i = 0; i < len; i++) {
-			items[i] = new RegExp(sources![i]!, flags)
-		}
-		rs.compiledItems = items
-	}
 
-	if (!items) return rs.list[rs.pattern]!
 	const len = items.length
 	for (let i = 0; i < len; i++) {
 		if (items[i]!.test(path)) return rs.list[i]!
 	}
-	return rs.list[rs.pattern]!
+	throw new Error("view-ignored has crashed: expected sub-pattern", {
+		cause: rs,
+	})
 }
 
 function getIgnoreOptions(options: RuleTestOptions, src: Resource): IgnoresOptions {
@@ -347,7 +338,7 @@ function getIgnoreOptions(options: RuleTestOptions, src: Resource): IgnoresOptio
 export function ruleTestSync(options: RuleTestOptions): RuleMatch {
 	const src = options.resource
 
-	if (src === undefined) throw new Error("view-ignored has crashed: no source cached.")
+	if (src === undefined) throw new Error("view-ignored has crashed: no source cached")
 
 	if (src !== null && "error" in src)
 		return { ...src, ignored: true, kind: RuleMatchKind.invalidSource }